Fixed app state bugs
- created a second app state isolated from BLE state
