+++ app/lib/app_state.dart
@@ -39,15 +39,5 @@ class AppState extends ChangeNotifier {
   
   void toggleWifi() {
     _wifiConnected = !_wifiConnected;
-  int pageIndex = 0;
-
-  void changeIndex(int ind){
-    pageIndex = ind;
-    notifyListeners();
-  }
-
-  void toggleConnection() {
-    isConnected = !isConnected;
-    notifyListeners();
   }
 }
