+++ app/lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
+import 'package:robot_app/app-state2.dart';
 
 import 'package:robot_app/app_state.dart';
 import 'app_layout.dart';
@@ -38,7 +39,7 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
  return ChangeNotifierProvider(
-      create: (_) => MyAppState(),
+      create: (_) => MyAppState1(),
       child: MaterialApp(
         title: 'Robot App',
